refactor: 사다리 게임 클래스 분리 및 각 클래스 역할에 맞게 메소드 이동, 메소드 명 변경
