+++ README.md
@@ -4,6 +4,20 @@
 
 ## 기능 요구사항
 
+### 사다리 게임
+- [x] 게임의 결과를 반환한다.
+  - [x] 플레이어의 보상을 반환한다.
+  - [x] 모든 플레이어의 보상을 반환한다.
+- [x] 보고 싶은 플레이어 결과를 반환한다.
+- [x] 모든 플레이어 결과를 반환한다.
+
+### 사다리
+- [x] 사다리 높이는 자연수이다.
+- [x] 사다리의 라인을 생성합니다.
+  - [x] 라인은 다리를 생성한다.
+  - [x] 다리는 랜덤으로 생성한다.
+  - [x] 라인을 이루는 다리가 겹치지 않도록 해야 한다.
+
 ### 플레이어
 - [x] 이름을 최대5글자까지 부여할 수 있다.
   - [x] ``예외`` : 5글자가 넘는 경우
@@ -13,15 +27,12 @@
 - [x] ``예외`` : 이름이 중복되는 경우
 - [x] 플레이어가 몇 번째 순서(index)인지 반환한다.(0부터시작)
 
-### 사다리
-- [x] 사다리 높이는 자연수이다.
-- [x] 사다리의 라인을 생성합니다.
-  - [x] 라인은 다리를 생성한다.
-  - [x] 다리는 랜덤으로 생성한다.
-  - [x] 라인을 이루는 다리가 겹치지 않도록 해야 한다.
-- [x] 플레이어의 보상을 반환한다.
-- [x] 모든 플레이어의 보상을 반환한다.
 
+### 보상
+- [x] ``예외`` : 보상이 5글자가 넘는 경우
+- [x] ``예외`` : 보상의 개수가 플레이어 수랑 같지 않을 경우
+- [x] ``예외`` : 빈 문자열인 경우
+- 
 ### 입력
 - [x] 플레이어 이름을 입력받는다.
   - [x] 이름은 쉼표(,)를 기준으로 구분한다.
@@ -31,7 +42,7 @@
   - [x] 모든 공백은 제거 한다.
 - [x] 사다리 높이를 입력받는다.
   - [x] ``예외`` : 숫자가 아닌 경우
-- [x] 결과를 보고 싶은 플레이어 입력받는다.
+- [ ] 결과를 보고싶은 플레이어를 입력받는다.
   - [x] all을 입력하기 전까지 계속 입력받는다.
   - [x] all을 입력할 경우 모두 출력하고 종료한다.
   - [x] 없는 플레이어를 입력할 경우 다시 입력받는다.
@@ -41,10 +52,3 @@
 - [x] 다리는 '-'를 5개 출력한다.
 - [x] 사다리 맨 앞 공백은 처음 사람 이름의 길이와 같다
 - [x] 보고 싶은 플레이어 결과 출력한다.
-
-
-### 보상
-- [x] ``예외`` : 보상이 5글자가 넘는 경우
-- [x] ``예외`` : 보상의 개수가 플레이어 수랑 같지 않을 경우
-- [x] ``예외`` : 빈 문자열인 경우
-
