+++ src/main/java/LadderApplication.java
@@ -1,9 +1,9 @@
-import controller.LadderGame;
+import controller.LadderController;
 
 public class LadderApplication {
 
     public static void main(String[] args) {
-        LadderGame ladderController = new LadderGame();
+        LadderController ladderController = new LadderController();
         ladderController.run();
     }
 }
