+++ src/main/java/domain/Leg.java
@@ -3,7 +3,6 @@ package domain;
 public class Leg {
     private final boolean isExistLeg;
 
-
     public Leg(boolean leg) {
         this.isExistLeg = leg;
     }
